+++ 2021_03_30_contacts/src/main/java/anna/demo/repo/ContactRepoMap.java
@@ -6,13 +6,14 @@ import org.springframework.stereotype.Repository;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 @Repository
 public class ContactRepoMap implements IContactRepo {
 
     HashMap<Integer, Contact> contactById = new HashMap<>();
-    private final AtomicLong idCounter = new AtomicLong(1);
+    private final AtomicInteger idCounter = new AtomicInteger(1);
 
     @Override
     public void save(Contact contact) {
@@ -41,6 +42,6 @@ public class ContactRepoMap implements IContactRepo {
     }
 
     public int createId(){
-        return (int) idCounter.getAndIncrement();
+        return idCounter.getAndIncrement();
     }
 }
